Moving classes to the correct folder
