+++ packages/lu/src/commands/qnamaker/convert.ts
@@ -4,10 +4,10 @@
  */
 
 import {CLIError, Command, flags, utils} from '@microsoft/bf-cli-command'
-const exception = require('./../../parser/lufile/classes/exception')
+const exception = require('./../../parser/utils/exception')
 const fs = require('fs-extra')
 const file = require('./../../utils/filehelper')
-const fileExtEnum = require('./../../parser/lufile/helpers').FileExtTypeEnum
+const fileExtEnum = require('./../../parser/utils/helpers').FileExtTypeEnum
 
 const QnAMaker = require('./../../parser/qna/qnamaker/qnamaker')
 const Alterations = require('./../../parser/qna/alterations/alterations')
