--- packages/lu/src/parser/lufile/classes/luObject.js
@@ -1,10 +1,0 @@
-
-class LuObject{
-    constructor(id, content, includeInCollate = true){
-        this.id = id
-        this.content = content
-        this.includeInCollate = includeInCollate
-    }
-}
-
-module.exports = LuObject;
