+++ packages/lu/src/parser/lufile/parseFileContents.js
@@ -3,29 +3,28 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-const LUISObjNameEnum = require('./enums/luisobjenum');
-const PARSERCONSTS = require('./enums/parserconsts');
-const builtInTypes = require('./enums/luisbuiltintypes');
-const helpers = require('./helpers');
+const LUISObjNameEnum = require('./../utils/enums/luisobjenum');
+const PARSERCONSTS = require('./../utils/enums/parserconsts');
+const builtInTypes = require('./../utils/enums/luisbuiltintypes');
+const helpers = require('./../utils/helpers');
 const chalk = require('chalk');
 const url = require('url');
-const retCode = require('./enums/CLI-errors');
+const retCode = require('./../utils/enums/CLI-errors');
 const parserObj = require('./classes/parserObject');
-const qnaListObj = require('./classes/qnaList');
-const qnaMetaDataObj = require('./classes/qnaMetaData');
+const qnaListObj = require('./../qna/qnamaker/qnaList');
+const qnaMetaDataObj = require('./../qna/qnamaker/qnaMetaData');
 const helperClass = require('./classes/hclasses');
 const deepEqual = require('deep-equal');
-const qna = require('./classes/qna');
-const exception = require('./classes/exception');
-const qnaAlterations = require('./classes/qnaAlterations');
+const exception = require('./../utils/exception');
+const qnaAlterations = require('./../qna/alterations/alterations');
 const fetch = require('node-fetch');
-const qnaFile = require('./classes/qnaFiles');
+const qnaFile = require('./../qna/qnamaker/qnaFiles');
 const fileToParse = require('./classes/filesToParse');
 const luParser = require('./luParser');
 const DiagnosticSeverity = require('./diagnostic').DiagnosticSeverity;
 const BuildDiagnostic = require('./diagnostic').BuildDiagnostic;
-const EntityTypeEnum = require('./enums/luisEntityTypes');
-const luisEntityTypeMap = require('./enums/luisEntityTypeNameMap');
+const EntityTypeEnum = require('./../utils/enums/luisEntityTypes');
+const luisEntityTypeMap = require('./../utils/enums/luisEntityTypeNameMap');
 const plAllowedTypes = ["composite", "ml"];
 const featureTypeEnum = {
     featureToModel: 'modelName',
@@ -1499,7 +1498,7 @@ const parseAndHandleEntity = function (parsedContent, luResource, log, locale) {
                     let alterationlist = [entity.Name];
                     if (entity.SynonymsOrPhraseList && entity.SynonymsOrPhraseList.length > 0) {
                         alterationlist = alterationlist.concat(entity.SynonymsOrPhraseList);
-                        parsedContent.qnaAlterations.wordAlterations.push(new qnaAlterations.alterations(alterationlist));
+                        parsedContent.qnaAlterations.wordAlterations.push(new qnaAlterations().wordAlterations = {"alterations": alterationlist});
                     } else {
                         let errorMsg = `QnA alteration section: "${alterationlist}" does not have list decoration. Prefix line with "-" or "+" or "*"`;
                         let error = BuildDiagnostic({
