+++ packages/lu/src/parser/utils/helpers.js
@@ -6,7 +6,7 @@
 const fs = require('fs');
 const path = require('path');
 const retCode = require('./enums/CLI-errors');
-const exception = require('./classes/exception');
+const exception = require('./exception');
 const NEWLINE = require('os').EOL;
 const ANY_NEWLINE = /\r\n|\r|\n/g;
 const url = require('url');
