+++ packages/lu/src/utils/filehelper.ts
@@ -6,20 +6,20 @@
 import {CLIError, utils} from '@microsoft/bf-cli-command'
 const fs = require('fs-extra')
 const path = require('path')
-const helpers = require('./../parser/lufile/helpers')
-const luObject = require('./../parser/lufile/classes/luObject')
+const helpers = require('./../parser/utils/helpers')
+const luObject = require('./../parser/lu/lu')
 
 /* tslint:disable:prefer-for-of no-unused*/
 
 export async function getLuObjects(stdin: string, input: string | undefined, recurse = false, extType: string | undefined) {
   let luObjects: any = []
   if (stdin) {
-    luObjects.push(new luObject('stdin', stdin))
+    luObjects.push(new luObject(stdin, 'stdin'))
   } else {
     let luFiles = await getLuFiles(input, recurse, extType)
     for (let i = 0; i < luFiles.length; i++) {
       let luContent = await getContentFromFile(luFiles[i])
-      luObjects.push(new luObject(path.resolve(luFiles[i]), luContent))
+      luObjects.push(new luObject(luContent, path.resolve(luFiles[i])))
     }
   }
