+++ packages/lu/test/parser/lufile/classes/classes.test.js
@@ -4,10 +4,10 @@
  */
 const chai = require('chai');
 const assert = chai.assert;
-const exception = require('./../../../../src/parser/lufile/classes/exception');
-const QnA = require('./../../../../src/parser/lufile/classes/qna');
-const QnAList = require('./../../../../src/parser/lufile/classes/qnaList');
-const QnAMetadata = require('./../../../../src/parser/lufile/classes/qnaMetaData');
+const exception = require('./../../../../src/parser/utils/exception');
+const QnA = require('./../../../../src/parser/qna/qnamaker/qnamaker');
+const QnAList = require('./../../../../src/parser/qna/qnamaker/qnaList');
+const QnAMetadata = require('./../../../../src/parser/qna/qnamaker/qnaMetaData');
 const hClasses = require('./../../../../src/parser/lufile/classes/hclasses');
 
 describe('Testing all classes', function() {
@@ -44,9 +44,10 @@ describe('Testing all classes', function() {
 
     describe('QnA class', function() {
         it('can create a new instance with explicit values as args', function() {
-            let urls = [];
             let qnatList = {test:123};
-            assert.deepEqual(new QnA(urls, qnatList).qnaList, qnatList);
+            let qna = new QnA()
+            qna.qnaList = qnatList
+            assert.deepEqual(qna.qnaList, qnatList)
         });
 
         it('can create a new instance with no values passed in', function() {
