+++ packages/lu/test/parser/lufile/parseFileContents.qnafile.test.js
@@ -6,7 +6,7 @@ const chai = require('chai');
 const assert = chai.assert;
 const parseFile = require('./../../../src/parser/lufile/parseFileContents').parseFile;
 const qnamaker = require('./../../../src/parser/qna/qnamaker/qnamaker')
-const retCode = require('./../../../src/parser/lufile/enums/CLI-errors').errorCode;
+const retCode = require('./../../../src/parser/utils/enums/CLI-errors').errorCode;
 describe('With parse file function', function() {
     it('Throws when input lu file has invalid URIs', function(done){
         let fileContent = `[InvalidPDF](https://download.microsoft.com/download/2/9/B/29B20383-302C-4517-A006-B0186F04BE28/surface-pro-4-user-guide-EN2.pdf)`;
